Moves loadscreens into LoadScreens folder/namespace. Moves LoadWidgetAtGameStart to Traits/World.
